v11

+++ src/accountdb/cardManagement.ts
@@ -138,6 +138,15 @@ export async function getCardTerms(cardId: number, cb: Function) {
 }
 
 export async function removeCardTerm(cardId: number, termId: number, cb: Function) {
+  let termExists = await prisma.term.findUnique({
+    where: {
+      id: termId
+    }
+  });
+  if (!termExists) {
+    console.log('Unknown term (double remove req?)');
+    return;
+  }
   await prisma.term.deleteMany({
     where: {
       id: termId,
@@ -146,6 +155,15 @@ export async function removeCardTerm(cardId: number, termId: number, cb: Functio
   });
 }
 export async function removeCard(cardId: number) {
+  let cardExists = await prisma.card.findUnique({
+    where: {
+      id: cardId
+    }
+  });
+  if (!cardExists) {
+    console.log('Unknown card (double remove req?)');
+    return;
+  }
   await prisma.term.deleteMany({
     where: {
       cardId
